Slight changes to Pipistrel
Change a few thngs to the pistrel, mainly input data. Also realized that I couldn't change some of the convergence parameters I setted so I  made some changes to the performance smodule from the pt file. For now it doesn't cause any problem because propellers are hte only propulsor, so it's safe to always promote "convergence:*" but it might cause problem later :bug:
Note: Pipistrel descent is made at 0kW
Also even though the e-motor of the pipistrel is very similar to the EMRAX 268, its efficiency is written as being equal to 89% vs 96% so we will add a parameter to adjust that in next commit