Voltage prediction for battery
As ever, the battery voltage was implemented as a particular case because wee need a precise prediction of its voltage to precisely predict the current (which we've seen in previous commit that a wrong prediction can prevent convergence). If the battery is directly connected to a bus, the battery becomes a voltage setter. Else, its voltage is simply set based on the number of cells in series. Next step is to set current