Cirrus SR22 integration tests
Made a proper folder for the SR22 and ensure that data were consistent with original model as this is one of the goals of this work. There is however a slight issue with input data which cause a useful load to be 20kg too much. This could be adjusted by reducing luggage mass from 55 to 35