Max power OPR limit done
Implemented the prediction of the max power due to the OPR limit. However, because what we look for in this fraework is the design point power, we had to "reverse" the original surrogate, which highlight the need for a different dataset hence why we built a different way of computing them. Surrogate is somewhat validated with a point known in the original framework