Mission test working
I had to do a very wonky thing to make the turboshaft model work. First, I introduced a new mechanic that allow to connect variable from performances to slipstream, but only if those variables are outputs of other components. Second I had to implement something that assumed that a variable will be connected from performances to slipstream and that will assume that this variable is bigger than anticipated and thuw we will cut it down. It's stupid but it works