Initial hybrid and Kodiak rework
Corrected a few input on the original Kodiak, mainly tto include the drag of strut and to have a better assessment of the impact on offtake (mechanical and bleed) on the engine performances
Also a first try of the hybridized Kodiak. Two main takes: battery are sized in C-rate and not in energy. And we could realistically use an EMRAX 348 for the hybridization provided we put a speed reducer. We should also try to see if there are batteries that have the same polarization curve and that can provide more c-rate even at the cost of energy