Hybrid improvement
Change a few things on the hybrid Kodiak, maily adjusted the mission parameters so they match the OP mission. Also took a value of cell fraction that corresponds to an existing aircraft and change the reference cell to one that provides more power than energy. However, this made me remark that the relative capacity was way too low for that cell as it has been computed based on a regression on another cell. This needs to be changed ! Maybe using options on the relative capacity component that provide a few points of the relative capacity curve ?