TBM rework
Based on the envisionned hybridation of the tbm we would need to add a propeller which is never used and so, to make it so that is still weighs something, we implemented a small mechanic (for now only on the propeller) to have the weight as an actual input. Used that mechanic to rework the TBM reference case. Does not change the results that much.