Retrofit of TBM
Made a first version of a TBM900 with 6 electric motor in a turbo electric configuration. Is very heavy and has a worst fuel consumption (as expected). It however made me realize there is a bug with the initilization of power. I currently don't have time to solve it but I added a test that displays it. This can be avoided by not preconditioning the powerrtain