EcoPulse + Hyrbid Kodiak rework
About finalized some corrections on the EcoPulse but the results doesn't seem too encouraging. Also rework a bit the Kodiak so that it uses the PW206B to the maxium it can not just its cruise power. This saves a bit of electric energy which in turns makes the PT lighter which almost enables us to carry the full max payload without crossing the MTOW