Propeller can now require 0 power
There was previously a convergence issue which forced the propeller shaf power to never be below a certain value (which the user could set), it can now go to 0 without convergence issue. But it raises an important point which is that many component "don't know" how to behave when the don't receive/produce power. Next on the line is the PMSM