Payload range computation working
The computation of the bounds of the payload range diagram is working. The associated post-processing function hasn't been coded yet nor has the computation of inner points (which should be fairly easy). On the electric version of the payload range point B and D are the same by convention (can't trade off fuel for PAX)