Inner payload range comparaison + sampling change
modified the plot for the inner payload range so that two aircraft data path can be given to allow the comparaison of the two. Also changed the sampling so that it covers more extensively the design space. Still need a slight tweek to make sure that point D is accounted for. Will see later