Final sampling change
In the unit tests is the use case that correspond to the comparaison of the Kodiak with its hybrid variant.
Also changed a little bit more the sampling to make sure to include point D in the data set which provide a cleaner payload range (although the bottom part and left part are still empty and will likely remain so).