First step
Using the lcav package, the goal seems to be able to automatically write a lca configuration file, the header consisting of the database and project title is fairly simple. For the methods, to get a proper rendering it is a bit mroe complicated but is done. Next step is to implement a way to do that components by components