Component level breakdown
A list of the impacts components by components seems to be working now. As mentioned in the comment of the code, however, the mechanism I implemented relies on the fact that the "component" custom attibutes (which is used to allocate impacts) should only be used when the impacts can be attributed to the component.