Airframe weight per FU
Added component to compute the weight of the airframe per fu will serve as the ref value to compute the impacts of the airframe. In this case we will consider that the airframe is one material (composite or aluminium), the flight contrls are in steel, the landing gears are in aluminium.