Tank "prod" + emissions per fu
Added the computation of the emissions per fu for the turboshaft and the ICE. Not yet accounted for in the impact. Also, I realized I had to put the tank weight per fu because, now, every compoentn needs a PreLCA group (which I'ven't added to the meta tests btw). Finally added a test case for a thermal propulsion chain.
Also Fixed a bug but I added a new one for balance.