Turboshaft operation
Added the necessary mechanism to write the combustion of kerozene inside the turboshaft. There is a bug where the phases are not properly computed because of lca_algebraix but it is solved by downloading the latest version of the master. This, however, can't be put in the pyproject.toml since lcav blocks the version of lca-algebraic. So it will need to be corrected later on.
Also added magic. Please do not touch