Use phase of battery (basically 0)
Added the stuff for the impact of the batteries. Is a bit rough tthe way it is done since we are adding process w know will be 0 but since they are in the conf file, they'll still appear as an additional line/colum in the matrix to invert.
This commit fixes everything else. Until the next commit.