Fuel production
Now included the production of the fuel (which includes the electricity for the battery) as a step in the LCA configuation file). I have doubts as the process i used as a proxy for the production of AvGas as the process doesn't exist in EcoInvent nor does the production of car gasoline.