Inventory for distribution
Computes the quantities necessary to evaluate the impact of the distribution phase. There will be two ways to distributes aircraft: either via the air (in which case we will have something very similar to the use and manufacturing phase) or via train.
However, because of refactoring, it is easier to compute the quantities necessary if the aircraft distributed via flight regardless of whether it is actually distributed via the air.