Equivalent midpoint weighting factor ReCiPe
Implemented equivalent weighting factors for ReCiPe to translate endpoints impact on midpoint indicators. As a check the value of the climate change impacts was compared and it gives roughly the same number. However, when looking at the proportion of each impact on the single score its not the same. Climate chnage account for way much and particulate matter formation as well. When looking at endpoint indicators human health seems to be predominant/