Disabling Recipe midpoints
Decided to backtrack and disable for now the use of the equivalent weighting factor to translate the single score to midpoint impact. The reason being it gave way too much weight to climate change and particulate matter formation. It might be due to how those are computed in ReCiPe. It needs further investigation