Possibility to use existing LCA conf file
Added the possibility to use existing LCA conf file and to specify the path to that file. It must however have the same format that the one generated by the code, mainly, the same emissions (even if nil) are expected. Tested it with the sensitivity analysis and appears to be working since the file that should have been overwritten if a new LCA file was generated was not modified.

Also fixed a bug which made it so that ionizing radiations were not considered