Rework of the assembly of some components
For a lot of components, we were originally using the data from an article but a second read of the paper made me realize that they might already account for the production of materials which I was also already counting. To avoid double counting and in absence of other sources, the assembly was discarded and only the prodcution of materials was considered.