Heavier Pipistrel
Small test case of what would happen if the composite structure of the Pipisrel was replaced by a metallic structure. Thus the aircraft is heavier, requires more energy, ... but the structure has a smaller enviornmental impact. Also included a buy to fly which is typically way more penalizing on metallic.