Test setup + firsy problem solved
A problem has been identified in which for powertrain with only a turboshaft, the model that computed the wing area based on the advanced aerodynamic model wouldn't work. This test reproduces it and a first step towards the solve is proposed. This test relies on reverting a changed I mad but can't rememember why