More warnings suppression
There still are a lot of warnings linked with divisions by zero in the np.where from the battery performances modules. Even though the np.where prevents nans and infinite from appearing in the OpenMDAO outputs, the Runtime warning still appears because the arrays with a zero divsion are still created although unused.