Implemented changes
Implemented the change so that now only what happens in the main route is considered. Thus, the correction I manually added can be removed. Several tests will be broken but fixed in the next commit.

One very important point is to know that the takeoff and climb won't be considered for impacts so future results might be slightly undesrtimated wwhen compared to current results with the correction