LCA considering different SOC start
First try at running the LCA straight after the sizing with differnet value for starting SOC leading to different depth of discharges and different lifetime. Initial results seem to indicate that it's not worth not fully discharging the battery. Was to be expected since 50%ish of the  battery is for reserve so DOD is always between 50 and 40% typically with very little effect accoridng to the author of the models