Test on Pipistrel
Tested the new model on the Pipistrel use case through the yml file which revealed a couple of things: as @felixpollet pointed out, you can't put ":" in option values in the yml so user is expected to put names with a "__" replacing the ":" the code then automatically replaces it. The code assess endurance that goes down very quickly with capacity much quicker that POH data. But then again, since SOH also affects the reserve then it should go down faster that SOH