Agign effect on energy per FU + sensitivity
Added a new model for the computation of the electricity per FU which allow the electricity per FU during operation to be an input since it was the results of private script I made. Also added a test which allows to evaluate the pipistrel with and without the full aging effect. Post processing to come