Added tests with nexw cell
Readded the baseline test on the Pipistrel including the LCA and added a new test with a different battery to see what effect it would have on the single score. This new battery which has high energy density and low lifecycle can significantly reduce the weight of the aircraft but at the cost of an increase single score. Detailed assumptions for the test are included in the test docstring