Added high lifespan case + small mods
Starting with the small modifcations: to ensure stability of the convergence process, the CG range for the aircraft is no longer computed rather it is an input whose value correspond to the reference case. Note that this means we are likely being conservative for the high lifespan case because batteries are heavier so the relative variation due to pax loading will likely e smaller.
Also added a new test case with batteries that are heavy but have very long lifespan (to the point where for a Pipistrel they wouldn't need to be replaced). Although these batteries have very low voltage (gor the high energy density case it was pretty similar to default case) so next commit will be about implementing a easier submodel