Adaptation to optionnal dependency
Adapted the code to handle the case where optionnal dependencies are not installed. I'm not really happy with the way it is done right now. What is happening when opt dependences are not added is:
- An user want to add the fastga_he.lca.legacy id to the conf file
- FAST-OAD tries to import the module where that id is defined but fails because of an import error
- the id is not recognized and the code stops here just saying that the id doesn't exist but without giving more details

I don't see an easy way of changing that, my opinion is that it's something that should be changed on FAST-OAD core

Also changed the workflow conf to not run certain tests linked with the environmental impacts if the version is not correct