Rework airspeed derivatives
For the computation of the airspeed derivation, significant imprecision were introduce because of the use of finit difference inside the `compute` plus it did not allow for analytical differentiation. It has been reworked based on a rewriting of the formula for the TAS as TAS = EAS * sqrt(rho_0/rho). Analytical diff has been added