"Fixed" DOE with power split
The power split exploration revealed that the variation of MTOW with power split prsented a weird behaviour. after investigation it was revealed that it was because of the instability of the electric motor model. Now it exhibits a much more reasonable shape.
Next commit will add a permanent fix for that weird behaviour